requirements.txt
fastapi~=0.67.0
numpy~=1.21.1
beautifulsoup4~=4.9.3
graypy~=2.1.0
sklearn
python-multipart
lxml==4.6.3
pydantic~=1.8.2
uvicorn
gunicorn
nltk~=3.6.2
PyYAML~=5.4.1
requests~=2.26.0
mongomock~=3.23.0
pymongo~=3.12.0
PyRSMQ
lightgbm~=3.2.1
redis~=3.5.3
aiofiles==0.6.0
pdfplumber
flake8
